// Day-29-CAPSTONE-INTEGRATE-POLISH/capstone-taskflow/src/app/models/auth.model.ts
export interface User {
  _id: string;
  name: string;
  email: string;
  role?: string;
  avatar?: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface RegisterRequest {
  name: string;
  email: string;
  password: string;
}

export interface AuthResponse {
  message: string;
  token: string;
  user: User;
}